feature/ssdExeRunner : [modify] interface와 object 파일 분리
