+++ SSD/TestShell/exeRunner.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <string>
+
+#define interface struct
+
+interface exeRunner {
+	virtual void runner(std::string cmd) = 0;
+};
